Move into kappa util, rename and document
