+++ indra/util/__init__.py
@@ -6,7 +6,6 @@ import gzip
 import zlib
 from io import BytesIO
 import xml.etree.ElementTree as ET
-from networkx import MultiDiGraph
 
 if sys.version_info[0] >= 3:
     non_unicode = bytes
@@ -200,32 +199,3 @@ def flatMap(f, xs):
     """Map a function onto an iterable and flatten the result."""
     return flatten(lmap(f, xs))
 
-def kappy_json_to_graph(kappy_json):
-    # handle the influence map case.
-    if 'influence map' in kappy_json.keys():
-        imap_data = kappy_json['influence map']['map']
-        graph = MultiDiGraph()
-
-        def add_edges(link_list, edge_sign):
-            for link_dict in link_list:
-                source = link_dict['source']
-                for target_dict in link_dict['target map']:
-                    target = target_dict['target']
-                    src_id = list(source.keys())[0] \
-                        + str(list(source.values())[0])
-                    tgt_id = list(target.keys())[0] \
-                        + str(list(target.values())[0])
-                    graph.add_edge(id_node_dict[src_id], id_node_dict[tgt_id],
-                                   sign = edge_sign)
-
-        id_node_dict = {}
-        for node_dict in imap_data['nodes']:
-            key = list(node_dict.keys())[0]
-            graph.add_node(node_dict[key]['label'], node_type=key)
-            new_key = key + str(node_dict[key]['id'])
-            id_node_dict[new_key] = node_dict[key]['label']
-
-        add_edges(imap_data['wake-up map'], 1)
-        add_edges(imap_data['inhibition map'], -1)
-
-    return graph
